Rename worker.go to proplet.go
